fix template sharing
